+++ crates/swc/tests/fixture/iissue-2118/input/index.js
@@ -1,0 +1,3 @@
+function test(a, b) {
+    a.__proto__ = b
+}