+++ crates/swc/tests/tsc-references/asyncArrowFunctionCapturesArguments_es5_es5.1.normal.js
@@ -61,21 +61,27 @@ var C = // @target: ES5
             value: function method() {
                 var other = function other() {
                 };
-                var fn = _asyncToGenerator(regeneratorRuntime.mark(function _callee() {
-                    var _args = arguments;
-                    return regeneratorRuntime.wrap(function _callee$(_ctx) {
-                        while(1)switch(_ctx.prev = _ctx.next){
-                            case 0:
-                                _ctx.next = 2;
-                                return other.apply(this, _args);
-                            case 2:
-                                return _ctx.abrupt("return", _ctx.sent);
-                            case 3:
-                            case "end":
-                                return _ctx.stop();
-                        }
-                    }, _callee, this);
-                }).bind(this)).bind(this);
+                var _this = this;
+                var fn = function() {
+                    var _ref = _asyncToGenerator(regeneratorRuntime.mark(function _callee() {
+                        var _args = arguments;
+                        return regeneratorRuntime.wrap(function _callee$(_ctx) {
+                            while(1)switch(_ctx.prev = _ctx.next){
+                                case 0:
+                                    _ctx.next = 2;
+                                    return other.apply(this, _args);
+                                case 2:
+                                    return _ctx.abrupt("return", _ctx.sent);
+                                case 3:
+                                case "end":
+                                    return _ctx.stop();
+                            }
+                        }, _callee, this);
+                    }).bind(_this)).bind(_this);
+                    return function fn() {
+                        return _ref.apply(this, arguments);
+                    };
+                }();
             }
         }
     ]);