+++ crates/swc/tests/tsc-references/computedPropertyNames8_ES6_es5.1.normal.js
@@ -11,11 +11,11 @@ function _defineProperty(obj, key, value) {
     }
     return obj;
 }
-var _obj;
 // @target: es6
 function f() {
     var t;
     var u;
+    var _obj;
     var v = (_obj = {
     }, _defineProperty(_obj, t, 0), _defineProperty(_obj, u, 1), _obj);
 }