+++ crates/swc/tests/vercel/full/react-instantsearch/2/output/index.js
@@ -8,49 +8,49 @@ import g from "./version";
 function h(a) {
     "function" == typeof a.addAlgoliaAgent && (a.addAlgoliaAgent("react (".concat(f, ")")), a.addAlgoliaAgent("react-instantsearch (".concat(g, ")")));
 }
-var i, j = function(a) {
+var i = function(a) {
     return e({
         ais: a.props.contextValue,
         multiIndexContext: a.props.indexContextValue
     });
-}, k = function(a, b) {
+}, j = function(a, b) {
     return a.props.indexContextValue.targetedIndex === b;
-}, l = function(a) {
+}, k = function(a) {
     return Boolean(a.props.indexId);
-}, m = function(a, b) {
+}, l = function(a, b) {
     return a.props.indexId === b;
-}, n = function(a, b) {
-    var c = l(a), d = l(b);
+}, m = function(a, b) {
+    var c = k(a), d = k(b);
     return c && !d ? -1 : !c && d ? 1 : 0;
 };
-export default function e(f) {
-    var g = f.indexName, p = f.initialState, q = f.searchClient, r = f.resultsState, s = f.stalledSearchDelay, t = function(a) {
-        return E.getWidgets().filter(function(a) {
+export default function n(e) {
+    var f = e.indexName, g = e.initialState, n = e.searchClient, p = e.resultsState, q = e.stalledSearchDelay, r = function(a) {
+        return C.getWidgets().filter(function(a) {
             return Boolean(a.getMetadata);
         }).map(function(b) {
             return b.getMetadata(a);
         });
-    }, u = function() {
-        var b = E.getWidgets().filter(function(a) {
+    }, s = function() {
+        var b = C.getWidgets().filter(function(a) {
             return Boolean(a.getSearchParameters);
         }).filter(function(a) {
-            return !j(a) && !l(a);
+            return !i(a) && !k(a);
         }).reduce(function(a, b) {
             return b.getSearchParameters(a);
-        }, D), c = E.getWidgets().filter(function(a) {
+        }, B), c = C.getWidgets().filter(function(a) {
             return Boolean(a.getSearchParameters);
         }).filter(function(a) {
-            var b = j(a) && k(a, g), c = l(a) && m(a, g);
+            var b = i(a) && j(a, f), c = k(a) && l(a, f);
             return b || c;
-        }).sort(n).reduce(function(a, b) {
+        }).sort(m).reduce(function(a, b) {
             return b.getSearchParameters(a);
-        }, b), d = E.getWidgets().filter(function(a) {
+        }, b), d = C.getWidgets().filter(function(a) {
             return Boolean(a.getSearchParameters);
         }).filter(function(a) {
-            var b = j(a) && !k(a, g), c = l(a) && !m(a, g);
+            var b = i(a) && !j(a, f), c = k(a) && !l(a, f);
             return b || c;
-        }).sort(n).reduce(function(b, c) {
-            var d = j(c) ? c.props.indexContextValue.targetedIndex : c.props.indexId, e = b[d] || [];
+        }).sort(m).reduce(function(b, c) {
+            var d = i(c) ? c.props.indexContextValue.targetedIndex : c.props.indexId, e = b[d] || [];
             return a.objectSpread({
             }, b, a.defineProperty({
             }, d, e.concat(c)));
@@ -67,35 +67,35 @@ export default function e(f) {
             mainParameters: c,
             derivedParameters: e
         };
-    }, v = function() {
-        if (!B) {
-            var a = u(A.state), b = a.mainParameters, c = a.derivedParameters;
-            A.derivedHelpers.slice().forEach(function(a) {
+    }, t = function() {
+        if (!z) {
+            var a = s(y.state), b = a.mainParameters, c = a.derivedParameters;
+            y.derivedHelpers.slice().forEach(function(a) {
                 a.detach();
             }), c.forEach(function(a) {
-                var b = a.indexId, c = a.parameters, d = A.derive(function() {
+                var b = a.indexId, c = a.parameters, d = y.derive(function() {
                     return c;
                 });
-                d.on("result", w({
+                d.on("result", u({
                     indexId: b
-                })).on("error", x);
-            }), A.setState(b), A.search();
+                })).on("error", v);
+            }), y.setState(b), y.search();
         }
-    }, w = function(b) {
+    }, u = function(b) {
         var c = b.indexId;
         return function(b) {
-            var d = I.getState(), e = !A.derivedHelpers.length, f = d.results ? d.results : {
+            var d = G.getState(), e = !y.derivedHelpers.length, f = d.results ? d.results : {
             };
             f = !e && f.getFacetByName ? {
             } : f, f = e ? b.results : a.objectSpread({
             }, f, a.defineProperty({
             }, c, b.results));
-            var g = I.getState(), h = g.isSearchStalled;
-            A.hasPendingRequests() || (clearTimeout(C), C = null, h = !1), g.resultsFacetValues;
+            var g = G.getState(), h = g.isSearchStalled;
+            y.hasPendingRequests() || (clearTimeout(A), A = null, h = !1), g.resultsFacetValues;
             var i = a.objectWithoutProperties(g, [
                 "resultsFacetValues"
             ]);
-            I.setState(a.objectSpread({
+            G.setState(a.objectSpread({
             }, i, {
                 results: f,
                 isSearchStalled: h,
@@ -103,19 +103,19 @@ export default function e(f) {
                 error: null
             }));
         };
-    }, x = function(b) {
-        var c = b.error, d = I.getState(), e = d.isSearchStalled;
-        A.hasPendingRequests() || (clearTimeout(C), e = !1), d.resultsFacetValues;
+    }, v = function(b) {
+        var c = b.error, d = G.getState(), e = d.isSearchStalled;
+        y.hasPendingRequests() || (clearTimeout(A), e = !1), d.resultsFacetValues;
         var f = a.objectWithoutProperties(d, [
             "resultsFacetValues"
         ]);
-        I.setState(a.objectSpread({
+        G.setState(a.objectSpread({
         }, f, {
             isSearchStalled: e,
             error: c,
             searching: !1
         }));
-    }, y = function(b, c) {
+    }, w = function(b, c) {
         if (b.transporter) {
             b.transporter.responsesCache.set({
                 method: "search",
@@ -153,7 +153,7 @@ export default function e(f) {
                 return a.concat(b.rawResults);
             }, [])
         })));
-    }, z = function(b, c) {
+    }, x = function(b, c) {
         if (b.transporter) {
             b.transporter.responsesCache.set({
                 method: "search",
@@ -183,28 +183,28 @@ export default function e(f) {
         }, d, JSON.stringify({
             results: c.rawResults
         })));
-    }, A = b(q, g, a.objectSpread({
+    }, y = b(n, f, a.objectSpread({
     }, d));
-    h(q), A.on("search", function() {
-        C || (C = setTimeout(function() {
-            var b = I.getState(), c = b.resultsFacetValues, d = a.objectWithoutProperties(b, [
+    h(n), y.on("search", function() {
+        A || (A = setTimeout(function() {
+            var b = G.getState(), c = b.resultsFacetValues, d = a.objectWithoutProperties(b, [
                 "resultsFacetValues"
             ]);
-            I.setState(a.objectSpread({
+            G.setState(a.objectSpread({
             }, d, {
                 isSearchStalled: !0
             }));
-        }, s));
-    }).on("result", w({
-        indexId: g
-    })).on("error", x);
-    var B = !1, C = null, D = A.state, E = c(function() {
-        var b = t(I.getState().widgets);
-        I.setState(a.objectSpread({
-        }, I.getState(), {
+        }, q));
+    }).on("result", u({
+        indexId: f
+    })).on("error", v);
+    var z = !1, A = null, B = y.state, C = c(function() {
+        var b = r(G.getState().widgets);
+        G.setState(a.objectSpread({
+        }, G.getState(), {
             metadata: b,
             searching: !0
-        })), v();
+        })), t();
     });
     !function(b, c) {
         if (c && (b.transporter && !b._cacheHydrated || b._useCache && "function" == typeof b.addAlgoliaAgent)) {
@@ -244,67 +244,68 @@ export default function e(f) {
                 };
             }
             if (Array.isArray(c.results)) {
-                y(b, c.results);
+                w(b, c.results);
                 return;
             }
-            z(b, c);
+            x(b, c);
         }
-    }(q, r);
-    var F, G, H, I = (G = {
-        widgets: void 0 === p ? {
-        } : p,
-        metadata: o(r),
-        results: (F = r) ? Array.isArray(F.results) ? F.results.reduce(function(c, d) {
+    }(n, p);
+    var D, E, F, G = (E = {
+        widgets: void 0 === g ? {
+        } : g,
+        metadata: o(p),
+        results: (D = p) ? Array.isArray(D.results) ? D.results.reduce(function(c, d) {
             return a.objectSpread({
             }, c, a.defineProperty({
             }, d._internalIndexId, new b.SearchResults(new b.SearchParameters(d.state), d.rawResults)));
         }, {
-        }) : new b.SearchResults(new b.SearchParameters(F.state), F.rawResults) : null,
+        }) : new b.SearchResults(new b.SearchParameters(D.state), D.rawResults) : null,
         error: null,
         searching: !1,
         isSearchStalled: !0,
         searchingForFacetValues: !1
-    }, H = [], {
+    }, F = [], {
         getState: function() {
-            return G;
+            return E;
         },
         setState: function(a) {
-            G = a, H.forEach(function(a) {
+            E = a, F.forEach(function(a) {
                 return a();
             });
         },
         subscribe: function(a) {
-            return H.push(a), function() {
-                H.splice(H.indexOf(a), 1);
+            return F.push(a), function() {
+                F.splice(F.indexOf(a), 1);
             };
         }
     });
     return {
-        store: I,
-        widgetsManager: E,
+        store: G,
+        widgetsManager: C,
         getWidgetsIds: function() {
-            return I.getState().metadata.reduce(function(a, b) {
+            return G.getState().metadata.reduce(function(a, b) {
                 return void 0 !== b.id ? a.concat(b.id) : a;
             }, []);
         },
-        getSearchParameters: u,
+        getSearchParameters: s,
         onSearchForFacetValues: function(b) {
             var c = b.facetName, d = b.query, e = b.maxFacetHits;
-            I.setState(a.objectSpread({
-            }, I.getState(), {
+            G.setState(a.objectSpread({
+            }, G.getState(), {
                 searchingForFacetValues: !0
-            })), A.searchForFacetValues(c, d, Math.max(1, Math.min(void 0 === e ? 10 : e, 100))).then(function(b) {
-                I.setState(a.objectSpread({
-                }, I.getState(), {
+            })), y.searchForFacetValues(c, d, Math.max(1, Math.min(void 0 === e ? 10 : e, 100))).then(function(b) {
+                var e;
+                G.setState(a.objectSpread({
+                }, G.getState(), {
                     error: null,
                     searchingForFacetValues: !1,
                     resultsFacetValues: a.objectSpread({
-                    }, I.getState().resultsFacetValues, (i = {
-                    }, a.defineProperty(i, c, b.facetHits), a.defineProperty(i, "query", d), i))
+                    }, G.getState().resultsFacetValues, (e = {
+                    }, a.defineProperty(e, c, b.facetHits), a.defineProperty(e, "query", d), e))
                 }));
             }, function(b) {
-                I.setState(a.objectSpread({
-                }, I.getState(), {
+                G.setState(a.objectSpread({
+                }, G.getState(), {
                     searchingForFacetValues: !1,
                     error: b
                 }));
@@ -315,33 +316,33 @@ export default function e(f) {
             });
         },
         onExternalStateUpdate: function(b) {
-            var c = t(b);
-            I.setState(a.objectSpread({
-            }, I.getState(), {
+            var c = r(b);
+            G.setState(a.objectSpread({
+            }, G.getState(), {
                 widgets: b,
                 metadata: c,
                 searching: !0
-            })), v();
+            })), t();
         },
         transitionState: function(a) {
-            var b = I.getState().widgets;
-            return E.getWidgets().filter(function(a) {
+            var b = G.getState().widgets;
+            return C.getWidgets().filter(function(a) {
                 return Boolean(a.transitionState);
             }).reduce(function(a, c) {
                 return c.transitionState(b, a);
             }, a);
         },
         updateClient: function(a) {
-            h(a), A.setClient(a), v();
+            h(a), y.setClient(a), t();
         },
         updateIndex: function(a) {
-            D = D.setIndex(a);
+            B = B.setIndex(a);
         },
         clearCache: function() {
-            A.clearCache(), v();
+            y.clearCache(), t();
         },
         skipSearch: function() {
-            B = !0;
+            z = !0;
         }
     };
 };