+++ crates/swc_ecma_preset_env/tests/fixtures/corejs3/usage-determanated-instance-methods/output.mjs
@@ -16,7 +16,7 @@ a.values();
 ''.includes();
 /./gi.flags;
 ''.indexOf();
-a["__defineGetter__"]('foo', function() {
+a.__defineGetter__('foo', function() {
     return 42;
 });
 a.trim;