+++ crates/swc_ecma_transforms_compat/tests/es2022_class_properties.rs
@@ -160,6 +160,31 @@ function (Bar) {
 
 test!(
     syntax(),
+    |_| chain!(
+        resolver(),
+        function_name(),
+        class_properties(class_properties::Config { loose: false }),
+    ),
+    private_class_method,
+    r#"
+class Foo {
+    #foo () {}
+}
+"#,
+    r#"
+var _foo = new WeakSet();
+class Foo {
+    constructor(){
+        _foo.add(this);
+    }
+}
+function foo() {
+}
+"#
+);
+
+test!(
+    syntax(),
     |t| tr(t),
     private_foobar,
     r#"
@@ -5234,6 +5259,45 @@ test!(
         }
     });
     "
+);
+
+test!(
+    syntax(),
+    |_| class_properties(class_properties::Config { loose: false }),
+    issue_3055_1,
+    "
+export class Node {
+    foo() {
+        this.#bar(this);
+    }
+
+    #bar(parent) {
+        parent.#baz(this);
+        parent.baz.#baz(this);
+    }
+
+    #baz(child) { }
+}
+",
+    "
+var _bar = new WeakSet(),
+    _baz = new WeakSet();
+export class Node {
+    foo() {
+        _classPrivateMethodGet(this, _bar, bar).call(this, this);
+    }
+    constructor() {
+        _bar.add(this);
+        _baz.add(this);
+    }
+}
+function bar(parent) {
+    var _baz1;
+    _classPrivateMethodGet(parent, _baz, baz).call(parent, this);
+    _classPrivateMethodGet(_baz1 = parent.baz, _baz, baz).call(_baz1, this);
+}
+function baz(child) {}
+"
 );
 
 test!(