+++ scripts/repo/profile-git.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/env bash
+#!/usr/bin/env zsh
 set -eu
 
 find . -type d -empty -delete